refactor code to use routes
